+++ com.unity.render-pipelines.high-definition/Editor/Lighting/Reflection/HDScreenSpaceReflectionEditor.cs
@@ -98,13 +98,11 @@ namespace UnityEditor.Rendering.HighDefinition
 
         void RayTracingQualityModeGUI()
         {
-            base.OnInspectorGUI();
-            using (new HDEditorUtils.IndentScope())
             using (new QualityScope(this))
             {
                 PropertyField(m_MinSmoothness, k_MinimumSmoothnessText);
                 PropertyField(m_SmoothnessFadeStart, k_SmoothnessFadeStartText);
-                m_SmoothnessFadeStart.value.floatValue = Mathf.Max(m_MinSmoothness.value.floatValue, m_SmoothnessFadeStart.value.floatValue);
+                m_SmoothnessFadeStart.value.floatValue  = Mathf.Max(m_MinSmoothness.value.floatValue, m_SmoothnessFadeStart.value.floatValue);
                 PropertyField(m_RayLength, k_RayLengthText);
                 PropertyField(m_ClampValue, k_ClampValueText);
                 PropertyField(m_SampleCount, k_SampleCountText);
@@ -290,6 +288,23 @@ namespace UnityEditor.Rendering.HighDefinition
             // SSR
             else
                 CopySetting(ref m_RayMaxIterations, settings.lightingQualitySettings.SSRMaxRaySteps[level]);
+        }
+
+        public override bool QualityEnabled()
+        {
+            // Quality always used for SSR
+            if (!HDRenderPipeline.rayTracingSupportedBySystem || !m_RayTracing.value.boolValue)
+                return true;
+
+            // Handle the quality usage for RTGI
+            HDRenderPipelineAsset currentAsset = HDRenderPipeline.currentAsset;
+
+            var bothSupportedAndPerformanceMode = currentAsset.currentPlatformRenderPipelineSettings.supportedRayTracingMode == RenderPipelineSettings.SupportedRayTracingMode.Both
+                && m_Mode.value.GetEnumValue<RayTracingMode>() == RayTracingMode.Performance;
+
+            var performanceSupported = currentAsset.currentPlatformRenderPipelineSettings.supportedRayTracingMode == RenderPipelineSettings.SupportedRayTracingMode.Performance;
+
+            return bothSupportedAndPerformanceMode || performanceSupported;
         }
     }
 }