+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -39,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed SSGI texture allocation when not using the RenderGraph.
 - Fixed NullReference Exception when setting Max Shadows On Screen to 0 in the HDRP asset.
 - Fixed path tracing accumulation not being reset when changing to a different frame of an animation.
+- Fixed issue with saving some quality settings in volume overrides  (case 1293747)
 
 ### Changed
 - Volume Manager now always tests scene culling masks. This was required to fix hybrid workflow.